Fake plugin 2.0

+++ plugins/fake.py
@@ -3,13 +3,14 @@ import os
 from pyrogram import Client, filters
 from pyrogram.raw import functions
 from pyrogram.errors import BadRequest, UsernameNotOccupied, UsernameInvalid, PeerIdInvalid
+from pyrogram.types import Message
 
 from db import db, save
 from config import cmds
 
 
 @Client.on_message(filters.command("fake", prefixes=".") & filters.me)
-async def fake(client, message):
+async def fake(client:Client, message:Message):
     text = message.text[6:]
     if message.reply_to_message:
         user_id = message.reply_to_message.from_user.id
@@ -29,12 +30,27 @@ async def fake(client, message):
     except BadRequest:
         await message.edit('only works with profiles.')
     else:
+        if not db['personal_data']['faked']:
+            info = await client.get_chat("me")
+            personal_data = dict(
+                first_name=info.first_name,
+                last_name=info.last_name or '',
+                description=info.bio or '',
+                faked=False,
+                user_photo=False
+            )
+
+            db['personal_data'] = personal_data
+            save(db)
+        if db['personal_data']['user_photo']:
+            photos =  await client.get_profile_photos("me")
+            await client.delete_profile_photos(photos[0].file_id)
+            db['personal_data']['user_photo'] = False
+            save(db)
         if cha.user.is_self:
             dat = db['personal_data']
-            try:
-                await client.set_profile_photo(photo='avatar.jpg')
-            except (BadRequest, FileNotFoundError):
-                pass
+            db['personal_data']['faked'] = False
+            save(db)
             text = 'No fake'
         else:
             text = 'new fake'
@@ -42,40 +58,25 @@ async def fake(client, message):
                 description = cha.about[:70]
             else:
                 description = ''
-            a = (await client.get_profile_photos(user_id, limit=1))[0]
+            db['personal_data']['faked'] = True
+            save(db)
+            if cha.user.photo:
+                a = (await client.get_profile_photos(user_id, limit=1))[0]
+                a = await client.download_media(a.file_id)
+                await client.set_profile_photo(photo=a)
+                os.remove(a)
+                db['personal_data']['user_photo'] = True
+                save(db)
             dat = dict(
                 first_name=cha.user.first_name,
                 last_name=cha.user.last_name or '',
                 description=description
             )
-            try:
-                a = await client.download_media(a.file_id)
-                await client.set_profile_photo(photo=a)
-                os.remove(a)
-            except:
-                pass
         await client.update_profile(first_name=dat['first_name'],
                                     last_name=dat['last_name'],
                                     bio=dat['description'])
         await message.edit(text)
+        print(db['personal_data'])
 
 
-@Client.on_message(filters.command("savepic", prefixes=".") & filters.me)
-async def savepic(client, message):
-    a = (await client.get_profile_photos("me", limit=1))[0]
-    try:
-        await client.download_media(a.file_id, a.file_ref, file_name='./avatar.jpg')
-        b = await client.get_chat("me")
-        personal_data = dict(
-            first_name=b.first_name,
-            last_name=b.last_name or '',
-            description=b.description or ''
-        )
-        db['personal_data'] = personal_data
-        save(db)
-        await message.edit('saved')
-    except Exception as e:
-        await message.edit(f'not saved\n\nCause: {e}')
-
-cmds.update({'.fake':"Copy the person's profile",
-             ".savepic":"Save your profile"})
+cmds.update({'.fake':"Copy the person's profile"})
